pyproject.toml
[tool.poetry]
name = "what-a-week-huh-bot"
version = "1.0.0"
description = ""
authors = ["Ondřej Hliba <hliba.ondra@live.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
pillow = "10.3.0"
telebot = "0.0.5"
schedule = "1.2.1"
pytelegrambotapi = "^4.18.1"
pytz = "2024.1"
pysoundfile = "0.9.0.post1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"